+++ forge/src/main/java/com/yanny/ali/plugin/mods/mantle/functions/package-info.java
@@ -1,0 +1,4 @@
+@ParametersAreNonnullByDefault
+package com.yanny.ali.plugin.mods.mantle.functions;
+
+import javax.annotation.ParametersAreNonnullByDefault;